+++ webview/src/components/SwaggerDocDrawer/DocCard/index.ts
@@ -1,0 +1,5 @@
+import DocCard from './DocCard';
+
+export default DocCard;
+
+export * from './DocCard';