+++ webview/src/components/SwaggerDocDrawer/context.ts
@@ -1,0 +1,8 @@
+import { SwaggerUrlConfigItem } from '@/utils/types';
+import React from 'react';
+
+export const SwaggerDocDrawerContext = React.createContext<{
+  swaggerUrlList: Partial<SwaggerUrlConfigItem>[];
+}>({ swaggerUrlList: [] });
+
+export const useSwaggerDocDrawerContext = () => React.useContext(SwaggerDocDrawerContext);