+++ webview/src/components/SwaggerDocDrawer/index.ts
@@ -1,0 +1,5 @@
+import SwaggerDocDrawer from './SwaggerDocDrawer';
+
+export default SwaggerDocDrawer;
+
+export * from './SwaggerDocDrawer';